Add the rest of the repositories

+++ domain/src/main/java/bartype/BarType.java
@@ -2,17 +2,17 @@ package bartype;
 
 public class BarType {
 
-    int id;
+    long id;
     private int beatCount; // e.g. for 6/8: 6
     private int beatValue; // e.g. for 6/8: 8
 
-    public BarType(int id, int beatCount, int beatValue) {
+    public BarType(long id, int beatCount, int beatValue) {
         this.id = id;
         this.beatCount = beatCount;
         this.beatValue = beatValue;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
